MAM: Add synchronous flag
This adds a synchronous access flag. It shrinks the maximum burst size
to 2^12. When the sync flag is set, the MAM issues req_sync with the
start of the burst. When the burst is complete it waits until the bus
interface asserts write_complete.

Currently the write_complete is tied to 1 for wishbone, because we
essentially only need a packet at the very end of the last burst to
confirm the anticipated completion of all writes. The later full
implementations should actually make sure that all outstanding writes
have completed. But for now the roundtrip time (the software block
until it receives the package) should be more than sufficient.