Interconnect: Add gateway router
With subnet support we're moving towards a world where HIM does not
represent the host as address 0 any more. Instead, traffic to the host is
forwarded by a gateway router (similar to a "standard gateway" in IP
networks).

The gateway router added in this commit has an additional output port
which receives all packets not intended for the local subnet (as
specified by the LOCAL_SUBNET parameter).