regaccess demux: Make ready more fine-grained
Make module work even if one of the out_bypass or out_reg ready signals
are permanently tied to zero. The module is now always ready as long as
it can fill its internal buffers until a packet header is received. Only
if the packet should be forwarded to either the reg or bypass output the
appropriate ready signal must be asserted.